Implemented thirst increase when in proximity to a well in the Geographer's planet desert biome. Work in progress though (see description)
It works! When you are close to a well it increases your thirst bar.
When you are not near a well, it decreases your thirst.

However, if your CurrentThirst level reaches the same as your MaxThirst
level, the HUD will not show the thirst decreasing!!

May need help debugging this one.

FYI: the MaxThirst and CurrentThirst are set in the EarthPrince
blueprint. (Ignore EarthPrince2 for now…that is my TESTING blueprint
haha!)